Style: Formatter code and add comments
